Common `LogMemStats` function
